clean up and consolidate
