+++ mod.ts
@@ -1,5 +1,4 @@
 export * from "./ops.ts";
-export * from "./ops-transitions.ts";
 export * from "./term.ts";
 export * from "./input.ts";
 export * from "./settings.ts";
